pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "the-ledger"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]",
    "asyncpg>=0.29.0",
    "openai>=1.0.0",
    "langgraph>=0.2.76",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "certifi",
    "pytest",
    "pytest-asyncio",
    "httpx",
    "uv"
]


[tool.pytest.ini_options]
asyncio_mode = "auto"
